metrics hello world
